# src/openapi.yaml
openapi: 3.1.3
info:
  title: Powercoders Chat API
  description: |-
    This is a samle chat api that can be used to create a Chat Frontend application

  contact:
    email: alex.trittibach@powercoders.org
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:3000
...
paths:
...
  /users:
    post:
      tags:
        - user
      summary: Create user
      description: Create your user
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: theUser
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: James
                password:
                  type: string
                  example: '12345'
      responses:
        201:
          description: Created
        400:
          description: Bad request
        401:
          description: Unauthorized


    get: 
      tags:
        - user
      summary: Get all users except of current
      description: This can only be done by the logged in user
      operationId: getUsers
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      security:
        - auth: []
...
components:
  schemas:
    User:
      type: object
      properties:
        authUserId:
          type: integer
          format: int64
          example: 10
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        lastActivity:
          type: string
          format: date-time
          example: '2023-12-14T14:57:21.508Z'
...
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT